+++ src/iotsaApiCoap.h
@@ -10,7 +10,7 @@ public:
   IotsaApiServiceCoap(IotsaApiProvider* _provider, IotsaApplication &_app, IotsaAuthenticationProvider* _auth, IotsaApiServiceProvider* _next=nullptr)
   : IotsaApiServiceCoap(_provider, _app, _next)
   {}
-  void setup(const char* path, bool get=false, bool put=false, bool post=false) override;
+  void setup(const char* path, bool get=false, bool put=false, bool post=false, bool webPage=true) override;
   static void ensureServiceMod(IotsaApplication &app);
 private:
   IotsaApiProvider* provider;